moved stub logger, model output equality unit tests pass, intergration tests are now at least collecting
